# Adds sampled flag to log correlation context