--- a/context/log4j12/src/test/java/brave/context/log4j12/MDCScopeDecoratorTest.java
+++ b/context/log4j12/src/test/java/brave/context/log4j12/MDCScopeDecoratorTest.java
@@ -73,12 +73,17 @@ public class MDCScopeDecoratorTest extends CurrentTraceContextTest {
           .isEqualTo(context.parentIdString());
       assertThat(MDC.get("spanId"))
           .isEqualTo(context.spanIdString());
+      Boolean sampled = context.sampled();
+      assertThat(MDC.get("sampled"))
+          .isEqualTo(sampled != null ? sampled.toString() : null);
     } else {
       assertThat(MDC.get("traceId"))
           .isNull();
       assertThat(MDC.get("parentId"))
           .isNull();
       assertThat(MDC.get("spanId"))
+          .isNull();
+      assertThat(MDC.get("sampled"))
           .isNull();
     }
   }