--- a/context/log4j2/README.md
+++ b/context/log4j2/README.md
@@ -15,7 +15,7 @@ tracing = Tracing.newBuilder()
     .build();
 ```
 
-Then, in your log configuration, you can use `traceId`, `parentId` and `spanId`.
+Then, in your log configuration, you can use `traceId`, `parentId`, `spanId` and `sampled`.
 
 Here's an example log4j2.properties pattern: