--- a/context/log4j2/src/test/java/brave/context/log4j2/ThreadContextScopeDecoratorTest.java
+++ b/context/log4j2/src/test/java/brave/context/log4j2/ThreadContextScopeDecoratorTest.java
@@ -45,12 +45,17 @@ public class ThreadContextScopeDecoratorTest extends CurrentTraceContextTest {
           .isEqualTo(context.parentIdString());
       assertThat(ThreadContext.get("spanId"))
           .isEqualTo(context.spanIdString());
+      Boolean sampled = context.sampled();
+      assertThat(ThreadContext.get("sampled"))
+          .isEqualTo(sampled != null ? sampled.toString() : null);
     } else {
       assertThat(ThreadContext.get("traceId"))
           .isNull();
       assertThat(ThreadContext.get("parentId"))
           .isNull();
       assertThat(ThreadContext.get("spanId"))
+          .isNull();
+      assertThat(ThreadContext.get("sampled"))
           .isNull();
     }
   }