Fixes #821