@adriancole 2018-12-11
ran benchmarks before and after.. nothing crazy