@adriancole 2018-12-11
oops missing one vote on #821 would anyone else use this?