@adriancole 2018-12-11
Here are the benchmarks. the results are unexpected from my pov (more keys more overhead), though #844 should help with this. Also later work can consider having the user whitelist keys they want to sync.

before

```
CurrentTraceContextBenchmarks.maybeScope_clear_default:maybeScope_clear_default·p0.999           sample             0.525             us/op
CurrentTraceContextBenchmarks.maybeScope_clear_default:·gc.alloc.rate.norm                       sample      15     0.003 ±   0.001    B/op
CurrentTraceContextBenchmarks.maybeScope_clear_log4j2:maybeScope_clear_log4j2·p0.999             sample            12.471             us/op
CurrentTraceContextBenchmarks.maybeScope_clear_log4j2:·gc.alloc.rate.norm                        sample      15  1392.152 ±   0.020    B/op
CurrentTraceContextBenchmarks.maybeScope_default:maybeScope_default·p0.999                       sample             0.794             us/op
CurrentTraceContextBenchmarks.maybeScope_default:·gc.alloc.rate.norm                             sample      15    24.005 ±   0.001    B/op
CurrentTraceContextBenchmarks.maybeScope_log4j2:maybeScope_log4j2·p0.999                         sample            14.362             us/op
CurrentTraceContextBenchmarks.maybeScope_log4j2:·gc.alloc.rate.norm                              sample      15  1680.199 ±   0.019    B/op
CurrentTraceContextBenchmarks.maybeScope_redundant_default:maybeScope_redundant_default·p0.999   sample             0.217             us/op
CurrentTraceContextBenchmarks.maybeScope_redundant_default:·gc.alloc.rate.norm                   sample      15     0.001 ±   0.001    B/op
CurrentTraceContextBenchmarks.maybeScope_redundant_log4j2:maybeScope_redundant_log4j2·p0.999     sample             0.251             us/op
CurrentTraceContextBenchmarks.maybeScope_redundant_log4j2:·gc.alloc.rate.norm                    sample      15     0.001 ±   0.001    B/op
CurrentTraceContextBenchmarks.newScope_clear_default:newScope_clear_default·p0.999               sample             0.214             us/op
CurrentTraceContextBenchmarks.newScope_clear_default:·gc.alloc.rate.norm                         sample      15     0.002 ±   0.001    B/op
CurrentTraceContextBenchmarks.newScope_clear_log4j2:newScope_clear_log4j2·p0.999                 sample            12.130             us/op
CurrentTraceContextBenchmarks.newScope_clear_log4j2:·gc.alloc.rate.norm                          sample      15  1392.144 ±   0.016    B/op
CurrentTraceContextBenchmarks.newScope_default:newScope_default·p0.999                           sample             0.272             us/op
CurrentTraceContextBenchmarks.newScope_default:·gc.alloc.rate.norm                               sample      15     0.002 ±   0.001    B/op
CurrentTraceContextBenchmarks.newScope_log4j2:newScope_log4j2·p0.999                             sample            13.845             us/op
CurrentTraceContextBenchmarks.newScope_log4j2:·gc.alloc.rate.norm                                sample      15  1680.182 ±   0.021    B/op
CurrentTraceContextBenchmarks.newScope_redundant_default:newScope_redundant_default·p0.999       sample             0.329             us/op
CurrentTraceContextBenchmarks.newScope_redundant_default:·gc.alloc.rate.norm                     sample      15     0.002 ±   0.001    B/op
CurrentTraceContextBenchmarks.newScope_redundant_log4j2:newScope_redundant_log4j2·p0.999         sample            11.232             us/op
CurrentTraceContextBenchmarks.newScope_redundant_log4j2:·gc.alloc.rate.norm                      sample      15  1048.132 ±   0.014    B/op
```

after

```
CurrentTraceContextBenchmarks.maybeScope_clear_default:maybeScope_clear_default·p0.999           sample             5.785             us/op
CurrentTraceContextBenchmarks.maybeScope_clear_default:·gc.alloc.rate.norm                       sample      15     0.003 ±   0.001    B/op
CurrentTraceContextBenchmarks.maybeScope_clear_log4j2:maybeScope_clear_log4j2·p0.999             sample            16.613             us/op
CurrentTraceContextBenchmarks.maybeScope_clear_log4j2:·gc.alloc.rate.norm                        sample      15  1728.201 ±   0.036    B/op
CurrentTraceContextBenchmarks.maybeScope_default:maybeScope_default·p0.999                       sample             2.632             us/op
CurrentTraceContextBenchmarks.maybeScope_default:·gc.alloc.rate.norm                             sample      15    24.007 ±   0.001    B/op
CurrentTraceContextBenchmarks.maybeScope_log4j2:maybeScope_log4j2·p0.999                         sample            20.850             us/op
CurrentTraceContextBenchmarks.maybeScope_log4j2:·gc.alloc.rate.norm                              sample      15  2232.314 ±   0.054    B/op
CurrentTraceContextBenchmarks.maybeScope_redundant_default:maybeScope_redundant_default·p0.999   sample             0.361             us/op
CurrentTraceContextBenchmarks.maybeScope_redundant_default:·gc.alloc.rate.norm                   sample      15     0.002 ±   0.001    B/op
CurrentTraceContextBenchmarks.maybeScope_redundant_log4j2:maybeScope_redundant_log4j2·p0.999     sample             2.723             us/op
CurrentTraceContextBenchmarks.maybeScope_redundant_log4j2:·gc.alloc.rate.norm                    sample      15     0.003 ±   0.001    B/op
CurrentTraceContextBenchmarks.newScope_clear_default:newScope_clear_default·p0.999               sample             0.481             us/op
CurrentTraceContextBenchmarks.newScope_clear_default:·gc.alloc.rate.norm                         sample      15     0.003 ±   0.001    B/op
CurrentTraceContextBenchmarks.newScope_clear_log4j2:newScope_clear_log4j2·p0.999                 sample            21.472             us/op
CurrentTraceContextBenchmarks.newScope_clear_log4j2:·gc.alloc.rate.norm                          sample      15  1728.283 ±   0.079    B/op
CurrentTraceContextBenchmarks.newScope_default:newScope_default·p0.999                           sample             0.402             us/op
CurrentTraceContextBenchmarks.newScope_default:·gc.alloc.rate.norm                               sample      15     0.003 ±   0.001    B/op
CurrentTraceContextBenchmarks.newScope_log4j2:newScope_log4j2·p0.999                             sample            25.306             us/op
CurrentTraceContextBenchmarks.newScope_log4j2:·gc.alloc.rate.norm                                sample      15  2232.349 ±   0.057    B/op
CurrentTraceContextBenchmarks.newScope_redundant_default:newScope_redundant_default·p0.999       sample             4.132             us/op
CurrentTraceContextBenchmarks.newScope_redundant_default:·gc.alloc.rate.norm                     sample      15     0.005 ±   0.002    B/op
CurrentTraceContextBenchmarks.newScope_redundant_log4j2:newScope_redundant_log4j2·p0.999         sample            11.716             us/op
CurrentTraceContextBenchmarks.newScope_redundant_log4j2:·gc.alloc.rate.norm                      sample      15  1552.238 ±   0.038    B/op
```