@adriancole 2018-12-11
hmm one issue is we haven't decided if "true" or "false" is the correct value for this, considering b3 is 1 or 0. I will re-open after we decide.