@b-devinssuresh-cvet 2019-03-01
Maybe put 2 fields: `sampled` and `sampledBool`?