@adriancole 2019-03-02
more fields is more overhead. I think probably sampled as boolean is best. if not present we clear the property. sound good?