@adriancole 2019-07-03
easiest is to use sampled as normal boolean as people can coerce the value to whatever they want in log config.